T210: Enable VPR TZ access
The VPR region is disabled in coreboot, but may need
to be reconfigured in the OS/drivers. To allow this,
we need to set the ALLOW_TZ_WRITE_ACCESS bit.

BUG=None
BRANCH=None
TEST=Built both Smaug and Foster OK.
Booted to kernel on my Smaug P5 OK. LCD showed 'Insert'
devmode info OK, but I don't have a FS on my eMMC so I
don't know if this change affects the kernel/OS UI.

Change-Id: I15e9f3cffffba0b89ed885b989dacae7b01ceaad
Signed-off-by: Tom Warren <twarren@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/275854
Reviewed-by: Aaron Durbin <adurbin@chromium.org>