DROP: uart8250mem: Fix bugs in MMIO uart handling
The handling for MMIO uarts is mostly non-functional if you are
not using the Oxford card.

This code has been reworked upstream so this patch should be
dropped when we rebase and we should not try to upstream it.

BUG=chrome-os-partner:40857
BRANCH=none
TEST=boot on glados with serial output

Change-Id: I092eac82cf10f3770bae4b1ecca685996cccf310
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/275158
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/275460
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>