skylake: Fix NVS C/ACPI synchronization
The C and ACPI NVS regions were not in sync.  Fix this and
remove the unused PRTx values.

I'm not sure what the RPAx is needed for either or why it is
in NVS to begin with.  These values are not being set yet and
the ACPI code needs it to be non-zero so this could still be
causing issues.

BUG=chrome-os-partner:40635
BRANCH=none
TEST=build and boot on glados

Change-Id: I4e806dc665707f0e2692ace6cdcd637abe7b8e4f
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/276344
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Aaron Durbin <adurbin@chromium.org>
Trybot-Ready: Aaron Durbin <adurbin@chromium.org>
Tested-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/276642
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>