glados: Support reading memory strap GPIOs to select SPD
Add some board specific code to enable the memory configuration
GPIOs in GPIO input mode and read them to determine which
memory type is on the board.

Also add the other memory types that are not yet present in
the glados mainboard directory.

This should be replaced with the real gpio infrastructure once
it is ready.

BUG=chrome-os-partner:43069
BRANCH=none
TEST=build and boot on glados

Change-Id: Iffb0bd5c176f2adbdd9302d9bff5b7bde725d671
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/287436
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/289770
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>