t210: lp0_resume: apply mbist WAR for audio on resume
When power is cut/restored to audio block, mbist workaround must be reapplied
or I2S will not function.  Handle this in lp0 resume firmware with the rest of
the mbist WAR.  This sequence for audio is also present in boot block code for
T210.

BUG=chrome-os-partner:41249
BRANCH=None
TEST=lp0 suspend/resume with audio playback

Change-Id: Ia6432e8556ee64f528d94f2dc3279b152294e132
Signed-off-by: Christopher Freeman <cfreeman@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/293618
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Anatol Pomazau <anatol@google.com>
Reviewed-by: Tom Warren <twarren@nvidia.com>
Reviewed-by: Anatol Pomazau <anatol@google.com>
Tested-by: Anatol Pomazau <anatol@google.com>