smaug: Reboot into EC RO if on recovery path
If device is booting in recovery path, ensure that EC is in RO
mode. This is to prevent depthcharge from doing a reboot very late
after it has realized that EC is not in RO.

Ensure that we do not lose the recovery reason. Also, if fastboot mode
is triggered manually, ensure that the recovery reason is properly
updated since we lose ec host events.

BUG=chrome-os-partner:31358
BRANCH=None
TEST=Compiles successfully and EC is put into RO before continuing on
recovery path

Change-Id: Ie5b2abd26761feaa932660bd0df24cd5d1be8ec7
Signed-off-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/295299
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>